First stab at establishing a mage public API
