+++ mage/BUILD
@@ -1,33 +1,3 @@
-load("@rules_cc//cc:defs.bzl", "cc_library")
-
-cc_library(
-  name = "mage",
-  hdrs = [
-    "bindings/receiver.h",
-    "bindings/remote.h",
-    "core/channel.h",
-    "core/core.h",
-    "core/endpoint.h",
-    "core/handles.h",
-    "core/message.h",
-    "core/node.h",
-    "core/util.h",
-  ],
-  srcs = [
-    "core/channel.cc",
-    "core/core.cc",
-    "core/endpoint.cc",
-    "core/message.cc",
-    "core/node.cc",
-    "core/util.cc",
-  ],
-  deps = [
-    "//base:base",
-    "@gtest//:gtest_prod",
-  ],
-  visibility = ["//visibility:public"],
-)
-
 # This filegroup target consists of misc binaries that //mage:mage_tests depend
 # on being built.
 filegroup(
@@ -57,7 +27,7 @@ config_setting(
 cc_test(
   name = "mage_tests",
   srcs = [
-    "core/mage_test.cc",
+    "//mage/core:mage_test.cc",
   ],
   copts = ["-Iexternal/gtest/include"],
   defines = select({
@@ -66,7 +36,8 @@ cc_test(
   }),
   deps = [
     "//base:base",
-    "//mage:mage",
+    "//mage/core",
+    "//mage/public:include",
     "//mage/test/magen:include",
     "@gtest//:gtest",
     "@gtest//:gtest_main",
