+++ mage/public/receiver.h
@@ -5,8 +5,8 @@
 
 #include "base/scheduling/scheduling_handles.h"
 #include "base/threading/thread_checker.h"
-#include "mage/core/handles.h"
-#include "mage/core/message.h"
+#include "mage/public/handles.h"
+#include "mage/public/message.h"
 
 namespace mage {
