+++ mage/public/util.h
