+++ mage/test/child_sends_send_invitation_pipe_to_parent.cc
@@ -10,9 +10,9 @@
 #include "base/scheduling/scheduling_handles.h"
 #include "base/scheduling/task_loop_for_io.h"
 #include "base/threading/thread_checker.h"
-#include "mage/bindings/remote.h"
-#include "mage/core/core.h"
-#include "mage/core/handles.h"
+#include "mage/public/remote.h"
+#include "mage/public/core.h"
+#include "mage/public/handles.h"
 #include "mage/test/magen/first_interface.magen.h"  // Generated.
 #include "mage/test/magen/second_interface.magen.h"  // Generated.
